Change expire_at to 64 bit signed int - [GP-479] (#9292)
* Change expire_at to 64 bit signed int

* Add test case for max expiry date of token

* Change name for max big int value constant

* Add comment on max expiry value

+++ database/migrations/2016_02_26_093844_create_tokens.php
@@ -103,7 +103,7 @@ class CreateTokens extends Migration
             $table->integer(Token::USED_AT)
                   ->nullable();
 
-            $table->integer(Token::EXPIRED_AT)
+            $table->bigInteger(Token::EXPIRED_AT)
                   ->nullable();
 
             $table->integer(Token::CREATED_AT);
